+++ hardpy/pytest_hardpy/db/schema/v1.py
@@ -217,6 +217,7 @@ class ResultStateStore(IBaseResult):
       "start_time": 1695817263,
       "status": "failed",
       "name": "hardpy-stand",
+      "alert": "",
       "dut": {
         "serial_number": "92c5a4bb-ecb0-42c5-89ac-e0caca0919fd",
         "part_number": "part_1",
@@ -303,6 +304,7 @@ class ResultStateStore(IBaseResult):
     dut: Dut
     modules: dict[str, ModuleStateStore] = {}
     operator_msg: dict = {}
+    alert: str
 
 
 class ResultRunStore(IBaseResult):