+++ hardpy/pytest_hardpy/reporter/hook_reporter.py
@@ -37,6 +37,7 @@ class HookReporter(BaseReporter):
         self.set_doc_value(DF.PROGRESS, 0, statestore_only=True)
         self.set_doc_value(DF.ARTIFACT, {}, runstore_only=True)
         self.set_doc_value(DF.OPERATOR_MSG, {}, statestore_only=True)
+        self.set_doc_value(DF.ALERT, "", statestore_only=True)
 
         test_stand_tz = self.generate_key(DF.TEST_STAND, DF.TIMEZONE)
         self.set_doc_value(test_stand_tz, str(get_localzone().key))
@@ -51,6 +52,7 @@ class HookReporter(BaseReporter):
         self.set_doc_value(DF.START_TIME, start_time)
         self.set_doc_value(DF.STATUS, TestStatus.RUN)
         self.set_doc_value(DF.PROGRESS, 0, statestore_only=True)
+        self.set_doc_value(DF.ALERT, "", statestore_only=True)
 
     def finish(self, status: TestStatus) -> None:
         """Finish test.
@@ -184,6 +186,14 @@ class HookReporter(BaseReporter):
             DF.ATTEMPT,
         )
         self.set_doc_value(key, attempt, statestore_only=True)
+
+    def set_alert(self, alert: str) -> None:
+        """Set alert message.
+
+        Args:
+            alert (str): alert message
+        """
+        self.set_doc_value(DF.ALERT, alert, statestore_only=True)
 
     def update_node_order(self, nodes: dict) -> None:
         """Update node order.